+++ src/platform_impl/windows/window.rs
@@ -188,7 +188,7 @@ impl Window {
                     | winuser::SWP_NOSIZE
                     | winuser::SWP_NOACTIVATE,
             );
-            winuser::UpdateWindow(self.window.0);
+            winuser::InvalidateRgn(self.window.0, ptr::null_mut(), 0);
         }
     }
 
@@ -507,7 +507,7 @@ impl Window {
                             size.1 as i32,
                             winuser::SWP_ASYNCWINDOWPOS | winuser::SWP_NOZORDER,
                         );
-                        winuser::UpdateWindow(window.0);
+                        winuser::InvalidateRgn(window.0, ptr::null_mut(), 0);
                     }
                 }
                 None => {
@@ -533,7 +533,7 @@ impl Window {
                                     | winuser::SWP_NOZORDER
                                     | winuser::SWP_NOACTIVATE,
                             );
-                            winuser::UpdateWindow(window.0);
+                            winuser::InvalidateRgn(window.0, ptr::null_mut(), 0);
                         }
                     }
                 }