+++ src/platform_impl/linux/x11/util/icon.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::window::{Icon, Pixel, PIXEL_SIZE};
+use crate::icon::{Icon, Pixel, PIXEL_SIZE};
 
 impl Pixel {
     pub fn to_packed_argb(&self) -> Cardinal {
@@ -18,13 +18,14 @@ impl Pixel {
 
 impl Icon {
     pub(crate) fn to_cardinals(&self) -> Vec<Cardinal> {
-        assert_eq!(self.rgba.len() % PIXEL_SIZE, 0);
-        let pixel_count = self.rgba.len() / PIXEL_SIZE;
-        assert_eq!(pixel_count, (self.width * self.height) as usize);
+        let rgba_icon = &self.inner;
+        assert_eq!(rgba_icon.rgba.len() % PIXEL_SIZE, 0);
+        let pixel_count = rgba_icon.rgba.len() / PIXEL_SIZE;
+        assert_eq!(pixel_count, (rgba_icon.width * rgba_icon.height) as usize);
         let mut data = Vec::with_capacity(pixel_count);
-        data.push(self.width as Cardinal);
-        data.push(self.height as Cardinal);
-        let pixels = self.rgba.as_ptr() as *const Pixel;
+        data.push(rgba_icon.width as Cardinal);
+        data.push(rgba_icon.height as Cardinal);
+        let pixels = rgba_icon.rgba.as_ptr() as *const Pixel;
         for pixel_index in 0..pixel_count {
             let pixel = unsafe { &*pixels.offset(pixel_index as isize) };
             data.push(pixel.to_packed_argb());