+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - On Windows, add `IconExtWindows` trait which exposes creating an `Icon` from an external file or embedded resource
 - Add `BadIcon::OsError` variant for when OS icon functionality fails
 - On Windows, fix crash at startup on systems that do not properly support Windows' Dark Mode
-- Use `i32` instead of `u32` for position type in `WindowEvent::Moved`.
+- **Breaking:** Use `i32` instead of `u32` for position type in `WindowEvent::Moved`.
 
 # 0.21.0 (2020-02-04)